pyproject.toml
[project]
name = "superstore-project"
version = "0.1.0"
description = ""
authors = [
    {name = "jinjik19",email = "ekrasnikov1918@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "pandas (==2.3.3)",
    "clickhouse-connect (==0.9.2)",
    "dbt-core (==1.10.13)",
    "dbt-clickhouse (==1.9.4)",
    "xlrd (==2.0.2)",
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
jupyterlab = "4.4.9"
apache-airflow = "3.1.0"
ruff = "0.14.0"
mypy = "1.18.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py312"
extend-exclude = ["__pycache__", "venv", "build", "dist"]
src = ["src", "dags"]

[tool.ruff.lint]
select = ["E", "F", "B", "W", "I"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["src", "dags"]
combine-as-imports = true
force-single-line = false
lines-after-imports = 2

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true
warn_return_any = true
warn_unused_ignores = true
strict_optional = true
no_implicit_optional = true
show_error_codes = true
pretty = true
exclude = ["venv", ".venv", "tests"]
files = ["src", "dags"]
